Simplify stream python implementation
    * Support '@stream' in documentation.
